+++ Assets/Scripts/Item.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 [Serializable]
 public class Item
 {
-    public int id;//唯一性ID
+    public long id;//唯一性ID
     public int itemId;//ItemTableData里面的ID
-    public int masterId;//拥有者
+    public long masterId;//拥有者
     public int itemLevel;
 }
